requirements.txt
pandas
torchvision
numpy
matplotlib
# reformer_pytorch